fixed all GPIO ordering, fully working on hardware
Fixed the column numbering which had been backwards, reordered sensorPin array to play low-to-high when strumming upwards. Code tested and working on handbuilt hardware over USB midi, have not yet tested DIN. Note that all changes to GPIOs was a result of the hardware used and not an issue with the original code.